+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/exception/GlobalExceptionHandler.java
@@ -12,7 +12,7 @@ import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 @RestControllerAdvice
-public class ControllerExceptionValidation extends ResponseEntityExceptionHandler {
+public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 
 	@ExceptionHandler(value = {MethodArgumentNotValidException.class})
 	public Map<String,Object> gestionValidaciones(MethodArgumentNotValidException ex) {
