+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/model/dto/EntradaDTO.java
@@ -1,4 +1,4 @@
-package com.recursosformacion.lcs.model_dto;
+package com.recursosformacion.lcs.model.dto;
 
 
 import org.springframework.stereotype.Component;
