+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/model/dto/NewUserForm.java
@@ -1,4 +1,4 @@
-package com.recursosformacion.lcs.model_dto;
+package com.recursosformacion.lcs.model.dto;
 
 import com.recursosformacion.lcs.util.constraint.interfaces.UserFormConstraint;
