+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/persistence/repository/ICine.java
@@ -1,8 +1,8 @@
-package com.recursosformacion.lcs.repository;
+package com.recursosformacion.lcs.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.recursosformacion.lcs.model.Cine;
+import com.recursosformacion.lcs.persistence.entity.Cine;
