+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/persistence/repository/IPrograma.java
@@ -1,9 +1,9 @@
-package com.recursosformacion.lcs.repository;
+package com.recursosformacion.lcs.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-import com.recursosformacion.lcs.model.Programa;
+import com.recursosformacion.lcs.persistence.entity.Programa;
