+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/service/IService.java
@@ -1,16 +1,13 @@
-package com.recursosformacion.lcs.service.interfaces;
+package com.recursosformacion.lcs.service;
 
 import java.util.List;
 import java.util.Optional;
-
-import org.springframework.stereotype.Service;
 
 import com.recursosformacion.lcs.exception.DAOException;
 import com.recursosformacion.lcs.exception.DomainException;
 
 
-@Service
-public interface  IServicio<T, S> {
+public interface IService<T, S> {
 
 	public T insert(T t) throws DomainException, DAOException ;
 	public boolean update(T t) throws DomainException, DAOException ;
