+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/service/impl/EntradaServiceImpl.java
@@ -1,24 +1,23 @@
-package com.recursosformacion.lcs.service;
+package com.recursosformacion.lcs.service.impl;
 
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.stereotype.Service;
-
 import com.recursosformacion.lcs.exception.DAOException;
 import com.recursosformacion.lcs.exception.DomainException;
-import com.recursosformacion.lcs.model.Entrada;
-import com.recursosformacion.lcs.repository.IEntrada;
-import com.recursosformacion.lcs.service.interfaces.IServicio;
+import com.recursosformacion.lcs.persistence.entity.Entrada;
+import com.recursosformacion.lcs.persistence.repository.IEntrada;
+import com.recursosformacion.lcs.service.IService;
 import com.recursosformacion.lcs.util.Rutinas;
+import org.springframework.stereotype.Service;
 
 @Service
-public class EntradaService implements IServicio<Entrada, Long> {
+public class EntradaServiceImpl implements IService<Entrada, Long> {
 
 	
 	private final IEntrada entradaRepository;
 	
-	EntradaService(IEntrada entradaRepository){
+	EntradaServiceImpl(IEntrada entradaRepository){
 		this.entradaRepository = entradaRepository;
 	}
